requirements.txt
discord.py
aiohttp
flask
gspread
requests
google-auth
oauth2client
asyncpg
aiosqlite
